Implementing client for upcoming movie

+++ MovieExplorerApi/MovieExplorerApi.Services/DTO/MovieResult.cs
@@ -13,7 +13,7 @@
         public string original_title { get; set; }
         public int[] genre_ids { get; set; }
         public string title { get; set; }
-        public int vote_average { get; set; }
+        public decimal vote_average { get; set; }
         public string overview { get; set; }
         public string release_date { get; set; }
     }
